chore(questions): clean up and comment
